+++ source/components/HOC/withApi.js
@@ -1,32 +1,59 @@
 // Core
 import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { bindActionCreators } from 'redux';
+import { connect } from 'react-redux';
 
-//api
-import { api, token } from "../../config/api";
+//bus
+import { tasksActions } from '../../bus/tasks/actions';
+import { postsActionsAsync } from '../../bus/tasks/saga/asyncActions';
 
-export const withApi = (Enchanced) =>
+export const withApi = (Enchanced) => {
+    const mapStateToProps = (state) => {
+        return {
+            isFetching: state.ui.get('isFetching'),
+            tasks:      state.tasks,
+        };
+    };
+
+    const mapDispathToProps = (dispatch) => {
+        return {
+            actions: bindActionCreators(
+                {
+                    ...tasksActions,
+                    ...postsActionsAsync,
+                },
+                dispatch
+            ),
+        };
+    };
+
+    @connect(mapStateToProps, mapDispathToProps)
     class WithApi extends Component {
 
-        constructor () {
-            super();
-            this.state = {
-                fetching: false,
-                tasks:    [],
-            };
-            this.fetchTask = ::this._fetchTask;
-            this.createTask = ::this._createTask;
-            this.changeTask = ::this._changeTask;
-            this.removeTask = ::this._removeTask;
-        }
+        static propTypes = {
+            actions:    PropTypes.object.isRequired,
+            isFetching: PropTypes.bool.isRequired,
+            tasks:      PropTypes.object.isRequired,
+        };
+
+        static defaultProps = {
+            isFetching: false,
+            tasks:      {},
+            actions:    {
+                fetchTasks:      () => {},
+                removeTaskAsync: () => {},
+                createTaskAsync: () => {},
+                changeTaskAsync: () => {},
+            },
+        };
 
         componentDidMount () {
-            //this.setOldTask();
-            //this.fetchTask();
-        }
+            const { actions } = this.props;
 
-        setFetching = (state) => {
-            this.setState({ fetching: state });
-        };
+            actions.fetchTasks();
+            //this.setOldTask();
+        }
 
         saveTaskLocal = (tasks) => {
             localStorage.setItem('tasks', JSON.stringify(tasks));
@@ -44,140 +71,13 @@ export const withApi = (Enchanced) =>
             }
         };
 
-        async _fetchTask () {
-            try {
-                this.setFetching(true);
-                const responce = await fetch(api, {
-                    method:  'GET',
-                    headers: {
-                        Authorization:  token,
-                        'Content-Type': 'application/json',
-                    },
-                });
-
-                if (responce.status !== 200) {
-                    throw new Error('Fetch failed');
-                }
-                const { data } = await responce.json();
-
-                this.setFetching(false);
-
-                this.setState(() => {
-                    const newTasks = [...data];
-
-                    this.saveTaskLocal(newTasks);
-
-                    return { tasks: newTasks };
-                });
-            } catch (error) {
-                this.setFetching(false);
-                console.error(error);
-            }
-        }
-
-        async _createTask (message) {
-            try {
-                this.setFetching(true);
-                const responce = await fetch(api, {
-                    method:  'POST',
-                    headers: {
-                        Authorization:  token,
-                        'Content-Type': 'application/json',
-                    },
-                    body: JSON.stringify({ message }),
-                });
-
-                if (responce.status !== 200) {
-                    throw new Error('Create failed');
-                }
-
-                const { data } = await responce.json();
-
-                this.setFetching(false);
-
-                this.setState(({ tasks }) => {
-                    const newTasks = [{ added: true, ...data }, ...tasks];
-
-                    this.saveTaskLocal(newTasks);
-
-                    return { tasks: newTasks };
-                });
-
-            } catch (error) {
-                this.setFetching(false);
-                console.error(error);
-            }
-        }
-
-        async _changeTask (task) {
-            try {
-                this.setFetching(true);
-                const responce = await fetch(api, {
-                    method:  'PUT',
-                    headers: {
-                        Authorization:  token,
-                        'Content-Type': 'application/json',
-                    },
-                    body: JSON.stringify(Array.isArray(task) ? task: [task]),
-                });
-
-                if (responce.status !== 200) {
-                    throw new Error('Change failed');
-                }
-
-                const { data } = await responce.json();
-
-                this.setFetching(false);
-
-                this.setState(({ tasks }) => {
-                    const newTasks = tasks.map((taskOld) => data[data.map((taskNew) => taskNew.id).indexOf(taskOld.id)] || taskOld);
-
-                    this.saveTaskLocal(newTasks);
-
-                    return { tasks: newTasks };
-                });
-            } catch (error) {
-                this.setFetching(false);
-                console.error(error);
-            }
-        }
-
-        async _removeTask (id) {
-            try {
-                this.setFetching(true);
-                const responce = await fetch(`${api}/${id}`, {
-                    method:  'DELETE',
-                    headers: {
-                        Authorization:  token,
-                        'Content-Type': 'application/json',
-                    },
-                });
-
-                if (responce.status !== 204) {
-                    throw new Error('Delete filed');
-                }
-
-                this.setFetching(false);
-
-                this.setState(({ tasks }) => {
-                    const newTasks = tasks.filter((task) => task.id !== id);
-
-                    this.saveTaskLocal(newTasks);
-
-                    return { tasks: newTasks };
-                });
-            } catch (error) {
-                this.setFetching(false);
-                console.error(error);
-            }
-        }
 
         render () {
             return (
-                <Enchanced
-                    { ...this.props }
-
-                />
+                <Enchanced { ...this.props } />
             );
         }
-    };
+    }
+
+    return WithApi;
+};
